+++ plugin/src/main/kotlin/com/chimerapps/storageinspector/ui/ide/InspectorSessionWindow.kt
@@ -35,6 +35,7 @@ import com.intellij.ui.content.Content
 import com.intellij.util.IconUtil
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.launch
 import java.awt.BorderLayout
 import java.io.File
@@ -62,7 +63,7 @@ class InspectorSessionWindow(
     private val serversView = StorageInspectorServersView(ProjectSessionIconProvider.instance(project), ::onServerSelectionChanged)
     private var currentDetailView: Any? = null
     private val splitter: JBSplitter
-    private val scope = CoroutineScope(Dispatchers.Default)
+    private val scope = CoroutineScope(SupervisorJob())
     private lateinit var connectActionGroup: DefaultActionGroup
 
     var connectionMode: ConnectionMode = ConnectionMode.MODE_DISCONNECTED